chore(turbo-tasks): Move Invalidator struct from manager.rs to invalidation.rs (#69073)
## Why?

`manager.rs` is getting *very* long (>2000 LOC) and hard to read
through. I've been only been making this worse in my stack. This PR is a
contribution to help offset this.

`Invalidator` seems like something that's very weakly coupled with the
rest of the manager, and there's a logical place for it in the
`invalidation` module.

## Testing

```
cargo nextest r -p turbo-tasks -p turbo-tasks-memory
```