refactor(turbo-tasks): Add stubs for RawVc::TaskOutput (#68908)
## Why?


https://www.notion.so/vercel/RawVc-LocalOutput-aede5f463f594ca58396eb3fdaffd865

When returning from a function using `local_cells`, we need to return a
`RawVc`, but we don't know that the output type is yet, so we need a
`RawVc` variant that does not require a type.

This is fundamentally the same problem that `RawVc::TaskOutput` solves.
`RawVc::LocalOutput` is the same thing, but for functions using
`local_cells`.

## Implementation

There's an (partially broken) implementation in #69126. This PR exists
to help break #69126 into more reviewable chunks.