Ensure externalized React entrypoints use vendored React in App Router (#69610)
In the App Router we usually use the bundler to resolve React to the vendored version. However, `react-dom/server` is not actually bundled but used as an external pointing to the vendored version. That means that the vendored version also needs to point to vendored React versions.

We already ensured this for some entrypoints (namely `react` and `scheduler`). However, React 19 started to import from `react-dom` from `react-dom` Client entrypoints so we need to start aliasing `react-dom` as from vendored React packages.

This wasn't a problem so far since the React version we support in Pages Router is the same we vendor. However, with https://github.com/vercel/next.js/pull/69484 we'll use a different React version in Pages Router i.e. vendored and installed React use different versions. In those setups, `react-dom/server.edge` in App Router was using the installed React version which led to an internals mismatch. 

There's no concrete test for this until https://github.com/vercel/next.js/pull/69484 lands. It should still land separately since it can theoretically lead to multiple instances of `react-dom` being loaded.

Technically, we'd need to apply the same alias we use during bundling.