Partial Prerendering Basepath Fix (#69520)
This adds new information to the routes manifests that is used during
deployment for pages rendered using Partial Prerendering (PPR). It
changes the route detection mode from using a pathname to detect a PPR
resume request to using a new header, `Next-Resume: 1`. When a `POST`
request is received with this header, it is assumed to be a resume
request.

The old mechanism is being retained until a future canary release to
validate the changes to infrastructure.