pages bugfix: prevent cancellation errors from being logged (#69766)
When another data request is received by the pages router, a
cancellation error is thrown and is meant to be caught by `change` to
signal that the navigation request was aborted. However we were logging
the error right before we handled the logic to properly bubble the error
up. This meant that it was logging an exception to the console that
wasn't actionable.

Fixes #40554
Closes NEXT-3731