Revert "Fix esm property def in flight loader (#66990)" (#69749)
This reverts commit 9bb06c546def5724a3a5235135681e6b3b130ba9.

This was a sort breaking change that could effect using `next/dynamic`
in server component, when the imported one is client component. Ideally
you shouldn't directly return it due to the property access restriction
for client reference. We've already provided related codemod for v15 but
this might be a bit surprise for users upgrading to the new patch.

Reverted as it's not necessarily part of ESM externals fixes.

Fixes #69720