Fix metadata prop merging (#69807)
### What

Fixes metadata merging issue on 14.2. There's a bug with React  on 14.2 as it bit outdated: while generating rsc payload, if we use fragment to wrap the metadata and noindex component, it will miss all the data. It works well with separating them as an array.
It also requires #64058 to be applied as well so the output not-found page will having correct status 404, and then we can generate noindex metatag based on correct status

Fixes #69778