Add support for `dynamicIO` experiment to `route.ts` modules
`dyanmicIO` enables a semantic where any uncached IO will be excluded from prerenders. Before PPR this means any uncached IO would bail out of static generation at build time and during revalidates. Route hanlders have a similar ability to be statically generated at build and revlidate time however we recently changed the behavior to only be enabled if you explicitly opt into this using an API like `export const revalidate` or `export const dynamic = "force-static"`. With dynamic IO though we don't really need to be this restrictive. For one those configs are going to be non-operational eventually with dynamicIO but additionally almost nothing would be static if it didn anything non-trivially static or if it was already opted into caching.

This commit updates the static generation rules for when `dynamicIO` is on to continue to try to statically generate GET route handlers. It also implements the logic for excluding IO and for when you use dynamic apis like `cookies()`.