remove unsafe-inline from with-strict-csp (#69497)
### What?
Remove unsafe-inline from with-strict-csp example in the script-src 

### Why?
The main purpose of having strict CSP headers is to improve the web app
security posture. On securityheaders.com the grade is capped to A if
script-src contains 'unsafe-inline'. In order to reach A+. it needs to
be removed.

### How?
Edit the script-src in the with-strict-csp-app example

Co-authored-by: JJ Kasper <jj@jjsweb.site>