Fix: (third-parties) sendGTMEvent not queueing events before GTM init (#68683)
<!-- Thanks for opening a PR! Your contribution is much appreciated.
To make sure your PR is handled as smoothly as possible we request that
you follow the checklist sections below.
Choose the right checklist for the change(s) that you're making:

## For Contributors

### Improving Documentation

- Run `pnpm prettier-fix` to fix formatting issues before opening the
PR.
- Read the Docs Contribution Guide to ensure your contribution follows
the docs guidelines:
https://nextjs.org/docs/community/contribution-guide

### Adding or Updating Examples

- The "examples guidelines" are followed from our contributing doc
https://github.com/vercel/next.js/blob/canary/contributing/examples/adding-examples.md
- Make sure the linting passes by running `pnpm build && pnpm lint`. See
https://github.com/vercel/next.js/blob/canary/contributing/repository/linting.md

### Fixing a bug

- Related issues linked using `fixes #number`
- Tests added. See:
https://github.com/vercel/next.js/blob/canary/contributing/core/testing.md#writing-tests-for-nextjs
- Errors have a helpful link attached, see
https://github.com/vercel/next.js/blob/canary/contributing.md

### Adding a feature

- Implements an existing feature request or RFC. Make sure the feature
request has been accepted for implementation before opening a PR. (A
discussion must be opened, see
https://github.com/vercel/next.js/discussions/new?category=ideas)
- Related issues/discussions are linked using `fixes #number`
- e2e tests added
(https://github.com/vercel/next.js/blob/canary/contributing/core/testing.md#writing-tests-for-nextjs)
- Documentation added
- Telemetry added. In case of a feature if it's used or not.
- Errors have a helpful link attached, see
https://github.com/vercel/next.js/blob/canary/contributing.md




## For Maintainers

- Minimal description (aim for explaining to someone not on the team to
understand the PR)
- When linking to a Slack thread, you might want to share details of the
conclusion
- Link both the Linear (Fixes NEXT-xxx) and the GitHub issues
- Add review comments if necessary to explain to the reviewer the logic
behind a change

### What?

sendGTMEvent pushes items to window.datalayer. THe GoogleTagManger (GTM)
client uses this to know what events to the GTM server.

currently if GTM is not ready, instead of adding events to the
dataLayer, it writes a warning to the console saying GTM is not ready.

The PR fixes that by defining the dataLayer if it is undefined so it can
start to queue up events before GTM is ready

### Why?

If GTM init is slow we don't want to loose any events that we want to
send before GTM init has happened we want to queue them up in the
dataLayer



### How?

The PR fixes that by defining the dataLayer if it is undefined so it can
start to queue up events before GTM is ready

Closes NEXT-
Fixes #

fixes #66865 (discussion)
fixes #68678 (issue)

-->

### What?

sendGTMEvent pushes items to window.datalayer. THe GoogleTagManger (GTM)
client uses this to know what events to the GTM server.

currently if GTM is not ready, instead of adding events to the
dataLayer, it writes a warning to the console saying GTM is not ready.

The PR fixes that by defining the dataLayer if it is undefined so it can
start to queue up events before GTM is ready


https://developers.google.com/tag-platform/tag-manager/datalayer#how_data_layer_information_is_processed

### Why?

If GTM init is slow we don't want to loose any events that we want to
send before GTM init has happened we want to queue them up in the
dataLayer

Another use case for this behaviour, is that you can optionally
initialise GoogleTagManager based on environment so you don't pollute
GTM data when developing or running playright tests, but you can still
evaluate what exists on the dataLayer to test if the correct events
would be sent to GTM after testing interactions with the app.

### How?

The PR fixes that by defining the dataLayer if it is undefined so it can
start to queue up events before GTM is ready

Fixes #

fixes #66865 (discussion)
fixes #68678 (issue)

---------

Co-authored-by: JJ Kasper <jj@jjsweb.site>