Add support for `dynamicIO` experiment to `route.ts` modules (#70208)
`dyanmicIO` enables a semantic where any uncached IO will be excluded
from prerenders. Before PPR this means any uncached IO would bail out of
static generation at build time and during revalidates.

Route handlers have a similar ability to be statically generated at
build and revalidate time however we recently changed the behavior to
only be enabled if you explicitly opt into this using an API like
`export const revalidate` or `export const dynamic = "force-static"`.
With dynamic IO though we don't really need to be this restrictive.

If you use `unstable_cache` or `fetch(..., { cache: 'force-cache' })`
you are opting into caching so when dynamicIO is on we can infer that
the GET handler is valid to cache as long as there is no uncached IO. So
in this commit we restore static generation when `dynamicIO` is enabled.