Remove dependecy graceful-fs and types (#67194)
Remove the dependency `graceful-fs` and types. 

I have looked at the project and it is only used in one file, inside
this file only `fs` is imported but not the functionality provided by
the library, finally it uses the native one through `graceful-fs`, I
have simplified the logic and removed the unnecessary library. The retry
functionality is implemented in the file and is not used from the
dependency

Co-authored-by: torresgol10.it <torresgol10.itd@gmail.com>
Co-authored-by: JJ Kasper <jj@jjsweb.site>