Avoid prefetch request before meta-tag-based redirect
When calling `redirect('/some-path')` in a server component after the
shell has already been served to the browser, we can no longer return a
response with a 307 status code. Instead, a meta tag is streamed into
the HTML document to trigger the navigation (e.g., `<meta http-equiv="refresh" content="1;url=/some-path" />`).

Since #63786, an additional client-side redirect in the client router is
suppressed when this meta tag is discovered. However, the fix did not
account for the ("temporary") prefetch request that may be triggered right
before the MPA navigation is executed.

With this PR, we now check for the meta tag before creating the prefetch
entry and return early if the tag is detected.