Support fallback params when the layout is a client component (#70329)
Stacked on #70328

When fallback params are passed to a layout we need to treat client
component layouts special because passing the fallback params to the
client component is otherwise enough to make them postpone. This change
introduces a similar abstraction that we use for client component Pages
where we pass our Page component to a PageRoot component which reifies
the dynamically tracked params from the underlying params.