feat(next-codemod): add `upgrade` command (#69935)
### Why?

Switching between `canary`, `rc`, and `latest` also requires React
version changes which could be time-consuming. This is essential for
reproduction, confirming issue fixed on canary, and migration. Also, to
run the codemods for breaking changes, users have to go through the
documentation and find which codemods to run for their project.

### How?

`@next/codemod upgrade` expects the user to select `canary`, `rc`, or
`latest` version to upgrade. Also, it detects any possible codemods to
run, and suggest the user whether to run it.


https://github.com/user-attachments/assets/3aa3fee8-75f2-4178-b5ce-fef725109d6b

---------

Co-authored-by: devjiwonchoi <devjiwonchoi@gmail.com>