feat(turbo-tasks): Accept ResolvedVc as a `self` argument (#70367)
#70269 adds support for `#[turbo_tasks::function]` to accept arguments
of type `ResolvedVc<...>`, while still exposing a signature that uses
`Vc<...>`.

This PR allows them to also accept `ResolvedVc<...>` for `self`, using
the nightly `arbitrary_self_types` feature.

This re-uses the same mechanisms used for other arguments, with the
added wrinkle that we can't shadow `self`, so we must rewrite the
function body to replace references to `self` with `turbo_tasks_self`.