feat(turbo-tasks): Ignore fields annotated with `#[turbo_tasks(trace_ignore)]` (#70375)
This annotation implies that the field does not contain any `Vc` types,
so we can re-use that here. We can skip checking for `ResolvedValue` on
these fields.

As we start codemodding stuff, we should consider renaming this
annotation to something more generic (e.g.
`#[turbo_tasks(does_not_contain_vc)]`), but this is good enough for now.