Avoid bundling react-dom/server.browser in Pages router
This was a regression introduced when extending support for Pages router
to React 18.
We have to keep the `try-catch` when the pages runtime is used as an external
(e.g. when the Node.js runtime is used).
However, when bundling, we can alias the module relevant to choosing the
correct `react-dom/server` entrypoint to the one we actually want.