add @taskr/esnext to pnpm root to resolve hoisting issues (#70349)
### What?

Add `@taskr/esnext` to the root package.json to get around hoisting
rules breaking `taskr`'s automatic discovery (and fix builds on some
platforms where it is broken).

### Why?

Some types of 'magically' imported packages (such as taskr automatically
supporting es6 by just adding a package, or drizzle automatically
finding a database driver) aren't hoisted on some platforms in the way
the tool expects, breaking it. The symptom of this is taskr throwing a
syntax error on some platforms, because we use ES6 syntax and it can't
find the package it needs to support it.

### How?

By adding it to the root package.json we ensure it is always hoisted in
the way taskr expects.