`dynamicIO` experiment without PPR should use canary React
Initially `dynamicIO` was implemented in a way that used only stable React APIs but we discovered an issue where you could observe something dynamic in SSR and possibly ignore it using try catch. React added an experimental feature for prerendering RSC which encodes incomplete chunks using a 'halt' mechanism that prevents the chunks from ever resolving. This allows the dynamic portions of the prerender to be unobservable from user code however to adopt this it meant that `dynamicIO` required experimental React. We don't generally want to require experimental React when using `dynamicIO` so this change updates our approach to go back to stable only APIs. The technique is to render RSC and SSR concurrently in the same task and then abort the SSR render in a following task. Because there is no PPR we can consider the result dynamic if either the RSC or SSR renders are incomplete. Since we abort the SSR render first we can also ensure that any non-microtasky chunks on the RSC side are pending before the SSR abort.

For synchronously dynamic APIs we might get a fully resolved RSC render before aborting the SSR render. In these cases it is possible in an edge case that a try catch is wrapped around a `use` which causes the sync error to be supressed in the SSR pass. This means we may not be able to accurate warn about the sync dynamci API however we'll still know the prerender was dynamic b/c we track this in the RSC render directly so we won't have any correctness issues, just interprettability ones. It will look like the case where there was something dynamic in RSC that was never observed by SSR which will be an error just not as specific as it could have been.